+++ semana10/labex/src/Common/Styles/Button.js
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+
+
+export const Button = styled.button`
+  border: none;
+  color: #ffffff;
+  background-color: #484CAD;
+  border-radius: 4px;
+  font-size: 1em;
+  padding: 8px;
+  box-shadow: 0 0 2px #111111;
+
+  :hover {
+    background-color: #5b60e5;
+    transition: 200ms
+  }
+`
